+++ CHANGELOG.md
@@ -13,6 +13,20 @@ The format is based on [Keep a Changelog][keepachangelog] and adheres to [Semant
 ### Deprecated
 
 ### Fixed
+## [0.69.4] - 2025-11-11
+
+### Added
+\n
+### Changed
+- RQLite database management is now integrated directly into the main node process, removing separate RQLite systemd services (debros-rqlite-*).
+- Improved log file provisioning to only create necessary log files based on the node type being installed (bootstrap or node).
+
+### Deprecated
+
+### Removed
+
+### Fixed
+\n
 ## [0.69.3] - 2025-11-11
 
 ### Added